feat: Criado componente Products para encapsular a responsabilidade de exibição de produtos.
