+++ src/app/lib/contants.ts
@@ -1,0 +1,3 @@
+const API_URL = 'http://localhost:3000';
+
+export { API_URL };
